v0.0.6
Change rating field data type to int64 in DB models

The data type of the 'Rating' field has been changed in the database models, queries, and SQL configurations, from 'int'/'INTEGER' to 'int64'/'BIGINT'. This change provides consistency between application code and the database schema, minimizing the potential for errors.

+++ austinapi_db/models.go
@@ -29,7 +29,7 @@ type Preparedness struct {
 type Sleep struct {
 	ID               int32
 	Date             time.Time
-	Rating           int
+	Rating           int64
 	TotalSleep       int
 	DeepSleep        int
 	LightSleep       int
